Simplify UiAutomator2 and XCUITest
