+++ src/main/java/id/aldochristiaan/salad/module/ios/Deeplink.java
@@ -1,10 +1,10 @@
 package id.aldochristiaan.salad.module.ios;
 
-import id.aldochristiaan.salad.module.Ios;
+import id.aldochristiaan.salad.module.XCUITest;
 import io.appium.java_client.ios.IOSDriver;
 import io.appium.java_client.ios.IOSElement;
 
-public class Deeplink extends Ios {
+public class Deeplink extends XCUITest {
 
     public Deeplink(IOSDriver<IOSElement> iosDriver) {
         super(iosDriver);
